An ingredient quantity can be a float
Like 1.5 spoon
@GautierMinster is tough on reviews

+++ model/recipe.go
@@ -1,7 +1,7 @@
 package model
 
 type Ingredient struct {
-	Quantity int
+	Quantity float64
 	Unit     string
 	Name     string
 	Category string
